Get the build working on MacOS
There were a bunch of changes here to get the C++ standard correct
because some of rapidyaml and it's dependencies didn't like the newer
standard.

I also removed the old cast warning for clang because there wasn't a way
to get c4core to build with the tests if it's there. Given that I'm not
going to try to fix c4core I figure this is ok.

The only Windows compatibility issue I'm worried about is '/' vs '\' in
path names for #include statements in the tests.